updated middleware to modularize the server
