+++ backend/controllers/queryController.ts
@@ -1,5 +1,9 @@
 const db = require('../models');
 
+const fetch = require('node-fetch');
+
+const { Pool } = require('pg');
+
 const queryController = {
   executeQueryUntracked: (req, res, next) => {
     // event.sender.send('async-started');
@@ -21,80 +25,46 @@ const queryController = {
       });
   },
 
-  executeQueryTracked: (req, res, next) => {
-    // create data object to send back to client
-    // const frontendData = {
-    //   queryData: null,
-    //   queryStats: null,
-    // };
-    // // extract query string from client request
-    // const { queryString } = req.body;
-    // // match the connection pool based on cookies
-    // const pool = users[req.cookies['session_id']];
-    // // retrieve query results and attach to frontend data object
-    // const rows = await pool.query(queryString);
-    // frontendData.queryData = rows;
-    // // Run EXPLAIN (FORMAT JSON, ANALYZE)
-    // if (!queryString.match(/create/i)) {
-    //   const queryStats = await pool.query(
-    //     'EXPLAIN (FORMAT JSON, ANALYZE) ' + queryString
-    //   );
-    //   frontendData.queryStats = queryStats;
-    // }
-    // // send back to client
-    // return res.status(200).json(frontendData);
+  executeQueryTracked: async (req, res, next) => {
+    // extract query string from client request
+    const { queryString, queryLabel } = req.body;
+    console.log('executeqeuery cookies', req.cookies);
+    // declare a user object to hold connection string
+    const users = {};
 
-    /*
-      event.sender.send('async-started');
-    
-      // destructure object from frontend
-      const { queryString, queryCurrentSchema, queryLabel } = data;
-    
-      // initialize object to store all data to send to frontend
-      let frontendData = {
-        queryString,
-        queryCurrentSchema,
-        queryLabel,
-        queryData: '',
-        queryStatistics: '',
-        lists: {},
-      };
-    
-      // Run select * from actors;
-      db.query(queryString)
-        .then((queryData) => {
-          frontendData.queryData = queryData.rows;
-          if (!queryString.match(/create/i)) {
-            // Run EXPLAIN (FORMAT JSON, ANALYZE)
-            db.query('EXPLAIN (FORMAT JSON, ANALYZE) ' + queryString).then(
-              (queryStats) => {
-                frontendData.queryStatistics = queryStats.rows;
-    
-                (async function getListAsync() {
-                  listObj = await db.getLists();
-                  frontendData.lists = listObj;
-                  event.sender.send('db-lists', listObj);
-                  event.sender.send('return-execute-query', frontendData);
-                  event.sender.send('async-complete');
-                })();
-              }
-            );
-          } else {
-            // Handling for tracking a create table query, can't run explain/analyze on create statements
-            (async function getListAsync() {
-              listObj = await db.getLists();
-              frontendData.lists = listObj;
-              event.sender.send('db-lists', listObj);
-              event.sender.send('async-complete');
-            })();
-          }
-        })
-        .catch((error: string) => {
-          console.log('ERROR in execute-query-tracked channel in main.ts', error);
-        });
-        */
+    const options = {
+      method: 'GET',
+      headers: {
+        Authorization:
+          'Basic Ojg4MDVmN2U2LTBiZWUtNDcwNC04OWRlLTU5YmM2ZTJlNWEyYw==',
+      },
+    };
+    const response = await fetch(
+      `https://customer.elephantsql.com/api/instances/${req.cookies.session_id}`,
+      options
+    );
+    const data = await response.json();
+    const { url } = data;
+    users[req.cookies.session_id] = new Pool({ connectionString: url });
 
-    next();
+    // match the connection pool based on cookies
+    const pool = users[req.cookies['session_id']];
+
+    // retrieve query results and attach to frontend data object
+    const rows = await pool.query(queryString);
+    res.locals.queryData = rows;
+
+    // Run EXPLAIN (FORMAT JSON, ANALYZE)
+    if (!queryString.match(/create/i)) {
+      const queryStats = await pool.query(
+        'EXPLAIN (FORMAT JSON, ANALYZE) ' + queryString
+      );
+      res.locals.queryStats = queryStats.rows;
+      res.locals.queryLabel = queryLabel;
+    }
+
+    // send back to client
+    return next();
   },
   generateDummyData: (req, res, next) => {
     next();
